Add end point to create User
This commit creates an endpoint 'POST /users/' to create a User.

This commit also arranges the User model and database connection
to a separate package named models and the function corresponding
to this point is stored in a separate package named 'controllers'.

This is an attempt to bring some similarity towards Ruby on Rails's
MVC approach.